Refactor stores, submithandler and checkboxline to common components at /NewConference/ (#80)
