+++ src/components/conference/NewConference/CheckboxLine.svelte
@@ -1,9 +1,9 @@
 <script>
     import FormField from "@smui/form-field";
     import Checkbox from "@smui/checkbox";
-    import {ConferenceCategory} from "../../../../enums/conference-category";
-    import {selectedCategoryTags} from "../newConferenceStores";
-    import JustifiedRow from "../../../form/JustifiedRow.svelte";
+    import {ConferenceCategory} from "../../../enums/conference-category";
+    import {selectedCategoryTags} from "./stores";
+    import JustifiedRow from "../../form/JustifiedRow.svelte";
 
     let categories = Object.keys(ConferenceCategory)
         .map(enumKey => ConferenceCategory[enumKey]);
