+++ src/components/conference/NewConference/Modal/DatePickerLine.svelte
@@ -1,7 +1,7 @@
 <script>
     import JustifiedRow from "../../../form/JustifiedRow.svelte";
     import DatePicker from "../../../form/DatePicker.svelte";
-    import {startDate, endDate, intervalWarning} from "../newConferenceStores.ts";
+    import {startDate, endDate, intervalWarning} from "../stores.ts";
