+++ src/components/conference/NewConference/Modal/Form.svelte
@@ -2,8 +2,8 @@
     import NameAndUrlLine from "./NameAndUrlLine.svelte";
     import DatePickerLine from "./DatePickerLine.svelte";
     import ButtonFooterLine from "./ButtonFooterLine.svelte";
-    import CheckboxLine from "./CheckboxLine.svelte";
-    import {pending} from "../newConferenceStores.ts";
+    import CheckboxLine from "../CheckboxLine.svelte";
+    import {pending} from "../stores.ts";
 
     export let formId;
 </script>
