+++ src/components/conference/NewConference/Page/Form.svelte
@@ -2,8 +2,8 @@
     import NameUrlLocationLine from "./NameUrlLocationLine.svelte";
     //import DatePickerLine from "./DatePickerLine.svelte";
     //import ButtonFooterLine from "./ButtonFooterLine.svelte";
-    //import CheckboxLine from "./CheckboxLine.svelte";
-    import {pending} from "../newConferenceStores.ts";
+    import CheckboxLine from "../CheckboxLine.svelte";
+    import {pending} from "../stores.ts";
 
     export let formId;
 </script>
